// src/plugins/core/types/match-mode.ts
export enum MatchMode {
  DateGte = 'DateGte',
  DateLte = 'DateLte',
  DateGt = 'DateGt',
  DateLt = 'DateLt',
  DateEqual = 'DateEqual',
  DateNotEqual = 'DateNotEqual',
  TextStarts = 'TextStarts',
  TextEnds = 'TextEnds',
  TextContains = 'TextContains',
  TextNotContains = 'TextNotContains',
  TextEqual = 'TextEqual',
  TextNotEqual = 'TextNotEqual',
  BoolTrue = 'BoolTrue',
  BoolFalse = 'BoolFalse',
  NumberGte = 'NumberGte',
  NumberGt = 'NumberGt',
  NumberLte = 'NumberLte',
  NumberLt = 'NumberLt',
  In = 'In',
  NotIn = 'NotIn',
  Between = 'Between',
  Exists = 'Exists',
}